+++ rows_test.go
@@ -14,306 +14,3 @@
 
 package spannerdriver
 
-import (
-	"context"
-	"database/sql"
-	"math"
-	"reflect"
-	"testing"
-)
-
-func TestRowsAtomicTypes(t *testing.T) {
-
-	// Open db.
-	ctx := context.Background()
-	db, err := sql.Open("spanner", dsn)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer db.Close()
-
-	// Set up test table.
-	_, err = db.ExecContext(ctx,
-		`CREATE TABLE TestAtomicTypes (
-			key	STRING(1024),
-			testString	STRING(1024),
-			testBytes	BYTES(1024),
-			testInt	INT64,
-			testFloat	FLOAT64,
-			testBool	BOOL
-		) PRIMARY KEY (key)`)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	_, err = db.ExecContext(ctx,
-		`INSERT INTO TestAtomicTypes (key, testString, testBytes, testInt, testFloat, testBool) VALUES
-		("general", "hello", CAST ("hello" as bytes), 42, 42, TRUE), 
-		("negative", "hello", CAST ("hello" as bytes), -42, -42, TRUE),
-		("maxint", "hello", CAST ("hello" as bytes), 9223372036854775807 , 42, TRUE),
-		("minint", "hello", CAST ("hello" as bytes), -9223372036854775808 , 42, TRUE),
-		("nan", "hello" ,CAST("hello" as bytes), 42, CAST("nan" AS FLOAT64), TRUE), 
-		("pinf", "hello" ,CAST("hello" as bytes), 42, CAST("inf" AS FLOAT64), TRUE),
-		("ninf", "hello" ,CAST("hello" as bytes), 42, CAST("-inf" AS FLOAT64), TRUE), 
-		("nullstring", null, CAST ("hello" as bytes), 42, 42, TRUE), 
-		("nullbytes", "hello", null, 42, 42, TRUE),
-		("nullint", "hello", CAST ("hello" as bytes), null, 42, TRUE), 
-		("nullfloat", "hello", CAST ("hello" as bytes), 42, null, TRUE), 
-		("nullbool", "hello", CAST ("hello" as bytes), 42, 42, null)
-		`)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	type testAtomicTypesRow struct {
-		key        string
-		testString string
-		testBytes  []byte
-		testInt    int
-		testFloat  float64
-		testBool   bool
-	}
-
-	type wantError struct {
-		scan  bool
-		close bool
-		query bool
-	}
-
-	tests := []struct {
-		name      string
-		input     string
-		want      []testAtomicTypesRow
-		wantError wantError
-	}{
-
-		{
-			name:  "general read",
-			input: `SELECT * FROM TestAtomicTypes WHERE key = "general"`,
-			want: []testAtomicTypesRow{
-				{key: "general", testString: "hello", testBytes: []byte("hello"), testInt: 42, testFloat: 42, testBool: true},
-			},
-		},
-		{
-			name:  "negative int and float",
-			input: `SELECT * FROM TestAtomicTypes WHERE key = "negative"`,
-			want: []testAtomicTypesRow{
-				{key: "negative", testString: "hello", testBytes: []byte("hello"), testInt: -42, testFloat: -42, testBool: true},
-			},
-		},
-		{
-			name:  "max int",
-			input: `SELECT * FROM TestAtomicTypes WHERE key = "maxint"`,
-			want: []testAtomicTypesRow{
-				{key: "maxint", testString: "hello", testBytes: []byte("hello"), testInt: 9223372036854775807, testFloat: 42, testBool: true},
-			},
-		},
-		{
-			name:  "min int",
-			input: `SELECT * FROM TestAtomicTypes WHERE key = "minint"`,
-			want: []testAtomicTypesRow{
-				{key: "minint", testString: "hello", testBytes: []byte("hello"), testInt: -9223372036854775808, testFloat: 42, testBool: true},
-			},
-		},
-		{
-			name:  "special float positive infinity",
-			input: `SELECT * FROM TestAtomicTypes WHERE key = "pinf"`,
-			want: []testAtomicTypesRow{
-				{key: "pinf", testString: "hello", testBytes: []byte("hello"), testInt: 42, testFloat: math.Inf(1), testBool: true},
-			},
-		},
-		{
-			name:  "special float negative infinity",
-			input: `SELECT * FROM TestAtomicTypes WHERE key = "ninf"`,
-			want: []testAtomicTypesRow{
-				{key: "ninf", testString: "hello", testBytes: []byte("hello"), testInt: 42, testFloat: math.Inf(-1), testBool: true},
-			},
-		},
-	}
-
-	// Run tests.
-	for _, tc := range tests {
-
-		rows, err := db.QueryContext(ctx, tc.input)
-		if (err != nil) && (!tc.wantError.query) {
-			t.Errorf("%s: unexpected query error: %v", tc.name, err)
-		}
-		if (err == nil) && (tc.wantError.query) {
-			t.Errorf("%s: expected query error but error was %v", tc.name, err)
-		}
-
-		got := []testAtomicTypesRow{}
-		for rows.Next() {
-			var curr testAtomicTypesRow
-			err := rows.Scan(
-				&curr.key, &curr.testString, &curr.testBytes, &curr.testInt, &curr.testFloat, &curr.testBool)
-			if (err != nil) && (!tc.wantError.scan) {
-				t.Errorf("%s: unexpected scan error: %v", tc.name, err)
-			}
-			if (err == nil) && (tc.wantError.scan) {
-				t.Errorf("%s: expected scan error but error was %v", tc.name, err)
-			}
-
-			got = append(got, curr)
-		}
-
-		rows.Close()
-		err = rows.Err()
-		if (err != nil) && (!tc.wantError.close) {
-			t.Errorf("%s: unexpected rows.Err error: %v", tc.name, err)
-		}
-		if (err == nil) && (tc.wantError.close) {
-			t.Errorf("%s: expected rows.Err error but error was %v", tc.name, err)
-		}
-
-		if !reflect.DeepEqual(tc.want, got) {
-			t.Errorf("Unexpected rows: %s. want: %v, got: %v", tc.name, tc.want, got)
-		}
-
-	}
-
-	// Drop table.
-	_, err = db.ExecContext(ctx, `DROP TABLE TestAtomicTypes`)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-}
-
-func TestRowsOverflowRead(t *testing.T) {
-
-	// Open db.
-	ctx := context.Background()
-	db, err := sql.Open("spanner", dsn)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer db.Close()
-
-	// Set up test table.
-	_, err = db.ExecContext(ctx,
-		`CREATE TABLE TestOverflowRead (
-			key	STRING(1024),
-			testString	STRING(1024),
-			testBytes	BYTES(1024),
-			testInt	INT64,
-			testFloat	FLOAT64,
-			testBool	BOOL
-		) PRIMARY KEY (key)`)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	_, err = db.ExecContext(ctx,
-		`INSERT INTO TestOverflowRead  (key, testString, testBytes, testInt, testFloat, testBool) VALUES
-		("general", "hello", CAST ("hello" as bytes), 42, 42, TRUE), 
-		("maxint", "hello", CAST ("hello" as bytes), 9223372036854775807 , 42, TRUE),
-		("minint", "hello", CAST ("hello" as bytes), -9223372036854775808 , 42, TRUE),
-		("max int8", "hello", CAST ("hello" as bytes), 127 , 42, TRUE),
-		("max uint8", "hello", CAST ("hello" as bytes), 255 , 42, TRUE)
-		`)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	type testAtomicTypesRowSmallInt struct {
-		key        string
-		testString string
-		testBytes  []byte
-		testInt    int8
-		testFloat  float64
-		testBool   bool
-	}
-
-	type wantError struct {
-		scan  bool
-		close bool
-		query bool
-	}
-
-	tests := []struct {
-		name      string
-		input     string
-		want      []testAtomicTypesRowSmallInt
-		wantError wantError
-	}{
-		{
-			name:  "read int64 into int8 ok",
-			input: `SELECT * FROM TestOverflowRead WHERE key = "general"`,
-			want: []testAtomicTypesRowSmallInt{
-				{key: "general", testString: "hello", testBytes: []byte("hello"), testInt: 42, testFloat: 42, testBool: true},
-			},
-		},
-		{
-			name:      "read int64 into int8 overflow",
-			input:     `SELECT * FROM TestOverflowRead WHERE key = "maxint"`,
-			wantError: wantError{scan: true},
-			want: []testAtomicTypesRowSmallInt{
-				{key: "maxint", testString: "hello", testBytes: []byte("hello"), testInt: 0, testFloat: 0, testBool: false},
-			},
-		},
-		{
-			name:  "read max int8 ",
-			input: `SELECT * FROM TestOverflowRead WHERE key = "max int8"`,
-			want: []testAtomicTypesRowSmallInt{
-				{key: "max int8", testString: "hello", testBytes: []byte("hello"), testInt: 127, testFloat: 42, testBool: true},
-			},
-		},
-		{
-			name:      "read max uint8 into int8",
-			input:     `SELECT * FROM TestOverflowRead WHERE key = "max uint8"`,
-			wantError: wantError{scan: true},
-			want: []testAtomicTypesRowSmallInt{
-				{key: "max uint8", testString: "hello", testBytes: []byte("hello"), testInt: 0, testFloat: 0, testBool: false},
-			},
-		},
-	}
-
-	// Run tests.
-	for _, tc := range tests {
-
-		rows, err := db.QueryContext(ctx, tc.input)
-		if (err != nil) && (!tc.wantError.query) {
-			t.Errorf("%s: unexpected query error: %v", tc.name, err)
-		}
-		if (err == nil) && (tc.wantError.query) {
-			t.Errorf("%s: expected query error but error was %v", tc.name, err)
-		}
-
-		got := []testAtomicTypesRowSmallInt{}
-		for rows.Next() {
-			var curr testAtomicTypesRowSmallInt
-			err := rows.Scan(
-				&curr.key, &curr.testString, &curr.testBytes, &curr.testInt, &curr.testFloat, &curr.testBool)
-			if (err != nil) && (!tc.wantError.scan) {
-				t.Errorf("%s: unexpected scan error: %v", tc.name, err)
-			}
-			if (err == nil) && (tc.wantError.scan) {
-				t.Errorf("%s: expected scan error but error was %v", tc.name, err)
-			}
-
-			got = append(got, curr)
-		}
-
-		rows.Close()
-		err = rows.Err()
-		if (err != nil) && (!tc.wantError.close) {
-			t.Errorf("%s: unexpected rows.Err error: %v", tc.name, err)
-		}
-		if (err == nil) && (tc.wantError.close) {
-			t.Errorf("%s: expected rows.Err error but error was %v", tc.name, err)
-		}
-
-		if !reflect.DeepEqual(tc.want, got) {
-			t.Errorf("%s: unexpected rows. want: %v, got: %v", tc.name, tc.want, got)
-		}
-
-	}
-
-	// Drop table.
-	_, err = db.ExecContext(ctx, `DROP TABLE TestOverflowRead`)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-}
